Resolved conflicts by switching all instances of Subexpression
with ScopedExpression.

Conflicts:
	src/edu/mit/compilers/optimization/AvailabilityCalculator.java

+++ src/edu/mit/compilers/tools/Dot.java
@@ -4,7 +4,7 @@ package edu.mit.compilers.tools;
 public class Dot {
     private Dot() {}
 
-    public static String node(int nodeId, String label) {
+    public static String node(long nodeId, String label) {
         // We prepend a backslash to any backslash or quote.
         // We need four backslashes to make it propogate through the Java String literal definition
         // and through the regex system.
@@ -13,7 +13,7 @@ public class Dot {
         return nodeId + "[label=\"" + escapedQuotes + "\"];";
     }
 
-    public static String edge(int sourceId, int destinationId) {
+    public static String edge(long sourceId, long destinationId) {
         return sourceId + " -> " + destinationId + ";";
     }
 }
